requirements.txt
pycryptodome>=3.20
cryptography>=42.0
